Split cookie flags checks rules
