+++ python/pyramid/audit/authtkt-cookie-httponly-unsafe-default.yaml
@@ -1,0 +1,28 @@
+rules:
+  - id: pyramid-authtkt-cookie-httponly-unsafe-default
+    patterns:
+      - pattern-either:
+          - patterns:
+              - pattern-not: pyramid.authentication.AuthTktCookieHelper(..., httponly=$HTTPONLY, ...)
+              - pattern-not: pyramid.authentication.AuthTktCookieHelper(..., **$PARAMS)
+              - pattern: pyramid.authentication.AuthTktCookieHelper(...)
+          - patterns:
+              - pattern-not: pyramid.authentication.AuthTktAuthenticationPolicy(..., httponly=$HTTPONLY, ...)
+              - pattern-not: pyramid.authentication.AuthTktAuthenticationPolicy(..., **$PARAMS)
+              - pattern: pyramid.authentication.AuthTktAuthenticationPolicy(...)
+    fix-regex:
+      regex: (.*)\)
+      replacement: \1, httponly=True)
+    message: >-
+      Found a Pyramid Authentication Ticket cookie without the httponly option correctly set. Pyramid
+      cookies should be handled securely by setting httponly=True.
+      If this parameter is not properly set, your cookies are not properly protected and
+      are at risk of being stolen by an attacker.
+    metadata:
+      cwe: "CWE-1004: Sensitive Cookie Without 'HttpOnly' Flag"
+      owasp: "A3: Sensitive Data Exposure"
+      category: security
+      technology:
+        - pyramid
+    languages: [python]
+    severity: WARNING
