+++ python/pyramid/audit/set-cookie-httponly-unsafe-value.yaml
@@ -1,5 +1,5 @@
 rules:
-  - id: pyramid-set-cookie-samesite
+  - id: pyramid-set-cookie-httponly-unsafe-value
     patterns:
       - pattern-either:
           - pattern-inside: |
@@ -13,18 +13,22 @@ rules:
                   ...
                   $RESPONSE = pyramid.httpexceptions.HTTPFound(...)
                   ...
-      - pattern-not: $RESPONSE.set_cookie(..., samesite='Lax', ...)
       - pattern-not: $RESPONSE.set_cookie(..., **$PARAMS)
-      - pattern-either:
-          - pattern: $RESPONSE.set_cookie(..., samesite=$SAMESITE, ...)
-          - pattern: $RESPONSE.set_cookie(...)
+      - pattern: $RESPONSE.set_cookie(..., httponly=$HTTPONLY, ...)
+      - pattern: $HTTPONLY
+      - metavariable-pattern:
+          metavariable: $HTTPONLY
+          pattern: |
+            False
+    fix: |
+      True
     message: >-
-      Found a Pyramid cookie without the samesite option correctly set. Pyramid
-      cookies should be handled securely by setting samesite='Lax' in response.set_cookie(...).
+      Found a Pyramid cookie without the httponly option correctly set.
+      Pyramid cookies should be handled securely by setting httponly=rue in response.set_cookie(...).
       If this parameter is not properly set, your cookies are not properly protected and
       are at risk of being stolen by an attacker.
     metadata:
-      cwe: "CWE-1275: Sensitive Cookie with Improper SameSite Attribute"
+      cwe: "CWE-1004: Sensitive Cookie Without 'HttpOnly' Flag"
       owasp: "A3: Sensitive Data Exposure"
       category: security
       technology:
