+++ python/pyramid/audit/set-cookie-samesite-unsafe-default.py
@@ -5,27 +5,9 @@ from pyramid.view import view_config
 
 
 @view_config(route_name='my_view')
-def my_bad_view1(request):
-    response = request.response
-    # ruleid: pyramid-set-cookie-samesite
-    response.set_cookie('MY_COOKIE', samesite=None,
-                        value='MY_COOKIE_VALUE', secure=True, httponly=True)
-    return {'foo': 'bar'}
-
-
-@view_config(route_name='my_view')
-def my_bad_view2(request):
-    resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ruleid: pyramid-set-cookie-samesite
-    resp.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
-                    samesite=None, secure=True, httponly=True)
-    return resp
-
-
-@view_config(route_name='my_view')
 def my_bad_view3(request):
     response = request.response
-    # ruleid: pyramid-set-cookie-samesite
+    # ruleid: pyramid-set-cookie-samesite-unsafe-default
     response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                         secure=True, httponly=True)
     return {'foo': 'bar'}
@@ -34,7 +16,7 @@ def my_bad_view3(request):
 @view_config(route_name='my_view')
 def my_bad_view4(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ruleid: pyramid-set-cookie-samesite
+    # ruleid: pyramid-set-cookie-samesite-unsafe-default
     resp.set_cookie('MY_COOKIE', secure=True, httponly=True)
     return resp
 
@@ -44,7 +26,7 @@ def my_bad_view4(request):
 @view_config(route_name='my_view')
 def my_good_view1(request):
     response = request.response
-    # ok: pyramid-set-cookie-samesite
+    # ok: pyramid-set-cookie-samesite-unsafe-default
     response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                         secure=True, httponly=True, samesite='Lax')
     return {'foo': 'bar'}
@@ -53,7 +35,7 @@ def my_good_view1(request):
 @view_config(route_name='my_view')
 def my_good_view3(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok: pyramid-set-cookie-samesite
+    # ok: pyramid-set-cookie-samesite-unsafe-default
     resp.set_cookie('MY_COOKIE', secure=True, httponly=True, samesite='Lax')
     return resp
 
@@ -61,6 +43,6 @@ def my_good_view3(request):
 @view_config(route_name='my_view')
 def my_good_view4(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok: pyramid-set-cookie-samesite
+    # ok: pyramid-set-cookie-samesite-unsafe-default
     resp.set_cookie('MY_COOKIE', **global_cookie_flags)
     return resp
