+++ python/pyramid/audit/set-cookie-samesite-unsafe-value.py
@@ -7,26 +7,18 @@ from pyramid.view import view_config
 @view_config(route_name='my_view')
 def my_bad_view1(request):
     response = request.response
-    # ruleid: pyramid-set-cookie-httponly
-    response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE', secure=True)
+    # ruleid: pyramid-set-cookie-samesite-unsafe-value
+    response.set_cookie('MY_COOKIE', samesite=None,
+                        value='MY_COOKIE_VALUE', secure=True, httponly=True)
     return {'foo': 'bar'}
 
 
 @view_config(route_name='my_view')
 def my_bad_view2(request):
-    response = request.response
-    # ruleid: pyramid-set-cookie-httponly
-    response.set_cookie('MY_COOKIE', httponly=False,
-                        value='MY_COOKIE_VALUE', secure=True)
-    return {'foo': 'bar'}
-
-
-@view_config(route_name='my_view')
-def my_bad_view3(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ruleid: pyramid-set-cookie-httponly
     resp.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
-                    secure=True)
+                    # ruleid: pyramid-set-cookie-samesite-unsafe-value
+                    samesite=None, secure=True, httponly=True)
     return resp
 
 
@@ -35,25 +27,23 @@ def my_bad_view3(request):
 @view_config(route_name='my_view')
 def my_good_view1(request):
     response = request.response
-    # ok: pyramid-set-cookie-httponly
+    # ok: pyramid-set-cookie-samesite-unsafe-value
     response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
-                        secure=True, httponly=True)
+                        secure=True, httponly=True, samesite='Lax')
     return {'foo': 'bar'}
-
-
-@view_config(route_name='my_view')
-def my_good_view2(request):
-    resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok: pyramid-set-cookie-httponly
-    resp.set_cookie('MY_COOKIE', secure=True,
-                    httponly=True, value='MY_COOKIE_VALUE')
-    return resp
 
 
 @view_config(route_name='my_view')
 def my_good_view3(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok: pyramid-set-cookie-httponly
-    resp.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
-                    **global_cookie_flags)
+    # ok: pyramid-set-cookie-samesite-unsafe-value
+    resp.set_cookie('MY_COOKIE', secure=True, httponly=True, samesite='Lax')
+    return resp
+
+
+@view_config(route_name='my_view')
+def my_good_view4(request):
+    resp = exc.HTTPFound(location=request.referer or request.application_url)
+    # ok: pyramid-set-cookie-samesite-unsafe-value
+    resp.set_cookie('MY_COOKIE', **global_cookie_flags)
     return resp
