+++ python/pyramid/audit/set-cookie-samesite-unsafe-value.yaml
@@ -1,5 +1,5 @@
 rules:
-  - id: pyramid-set-cookie-httponly
+  - id: pyramid-set-cookie-samesite-unsafe-value
     patterns:
       - pattern-either:
           - pattern-inside: |
@@ -13,19 +13,21 @@ rules:
                   ...
                   $RESPONSE = pyramid.httpexceptions.HTTPFound(...)
                   ...
-      - pattern-either:
-          - pattern: $RESPONSE.set_cookie(..., httponly=False, ...)
-          - patterns:
-              - pattern-not: $RESPONSE.set_cookie(..., httponly=True, ...)
-              - pattern-not: $RESPONSE.set_cookie(..., **$PARAMS)
-              - pattern: $RESPONSE.set_cookie(...)
+      - pattern-not: $RESPONSE.set_cookie(..., **$PARAMS)
+      - pattern: $RESPONSE.set_cookie(..., samesite=$SAMESITE, ...)
+      - pattern: $SAMESITE
+      - metavariable-regex:
+          metavariable: $SAMESITE
+          regex: (?!'Lax')
+    fix: |
+      'Lax'
     message: >-
-      Found a Pyramid cookie without the httponly option correctly set. Pyramid
-      cookies should be handled securely by setting httponly=True in response.set_cookie(...).
+      Found a Pyramid cookie without the samesite option correctly set.
+      Pyramid cookies should be handled securely by setting samesite='Lax' in response.set_cookie(...).
       If this parameter is not properly set, your cookies are not properly protected and
       are at risk of being stolen by an attacker.
     metadata:
-      cwe: "CWE-1004: Sensitive Cookie Without 'HttpOnly' Flag"
+      cwe: "CWE-1275: Sensitive Cookie with Improper SameSite Attribute"
       owasp: "A3: Sensitive Data Exposure"
       category: security
       technology:
