+++ python/pyramid/audit/set-cookie-secure-unsafe-default.py
@@ -7,7 +7,7 @@ from pyramid.view import view_config
 @view_config(route_name='my_view')
 def my_bad_view1(request):
     response = request.response
-    # ruleid: pyramid-set-cookie-secure
+    # ruleid: pyramid-set-cookie-secure-unsafe-default
     response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                         httponly=True)
     return {'foo': 'bar'}
@@ -15,17 +15,8 @@ def my_bad_view1(request):
 
 @view_config(route_name='my_view')
 def my_bad_view2(request):
-    response = request.response
-    # ruleid: pyramid-set-cookie-secure
-    response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
-                        secure=False, httponly=True)
-    return {'foo': 'bar'}
-
-
-@view_config(route_name='my_view')
-def my_bad_view3(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ruleid: pyramid-set-cookie-secure
+    # ruleid: pyramid-set-cookie-secure-unsafe-default
     resp.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                     httponly=True)
     return resp
@@ -36,7 +27,7 @@ def my_bad_view3(request):
 @view_config(route_name='my_view')
 def my_good_view1(request):
     response = request.response
-    # ok:  pyramid-set-cookie-secure
+    # ok:  pyramid-set-cookie-secure-unsafe-default
     response.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                         secure=True, httponly=True)
     return {'foo': 'bar'}
@@ -45,7 +36,7 @@ def my_good_view1(request):
 @view_config(route_name='my_view')
 def my_good_view2(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok:  pyramid-set-cookie-secure
+    # ok:  pyramid-set-cookie-secure-unsafe-default
     resp.set_cookie('MY_COOKIE', secure=True, httponly=True,
                     value='MY_COOKIE_VALUE')
     return resp
@@ -54,7 +45,7 @@ def my_good_view2(request):
 @view_config(route_name='my_view')
 def my_good_view3(request):
     resp = exc.HTTPFound(location=request.referer or request.application_url)
-    # ok:  pyramid-set-cookie-secure
+    # ok:  pyramid-set-cookie-secure-unsafe-default
     resp.set_cookie('MY_COOKIE', value='MY_COOKIE_VALUE',
                     **global_cookie_flags)
     return resp
